pyproject.toml
[project]
name = "triproxim8-py"
version = "0.1.0"
description = "Approximating images in low resolution with a genetic algorithm"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pygame>=2.6.0",
  "PyGLM>=2.8.0",
]

[project.scripts]
main = "triproxim8_py.app:main"
triproxim8 = "triproxim8_py.app:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/triproxim8_py"]
